+++ app/services/stellar_base/account_subscriptions/save_cursor.rb
@@ -1,12 +1,14 @@
 module StellarBase
-  class SaveCursor
+  module AccountSubscriptions
+    class SaveCursor
 
-    extend LightService::Action
-    expects :account_subscription, :operations
+      extend LightService::Action
+      expects :account_subscription, :operations
 
-    executed do |c|
-      c.account_subscription.update_attributes!(cursor: c.operations.last.id)
+      executed do |c|
+        c.account_subscription.update!(cursor: c.operations.last.id)
+      end
+
     end
-
   end
 end