+++ spec/dummy/app/services/get_how.rb
@@ -1,8 +1,8 @@
 class GetHow
 
-  SAMPLE_BTC_ADDRESS = "1BvBMSEYstWetqTFn5Au4m4GFg7xJaNVN2k"
+  SAMPLE_BTC_ADDRESS = "1BvBMSEYstWetqTFn5Au4m4GFg7xJaNVN2k".freeze
 
-  def self.call
+  def self.call(params)
     SAMPLE_BTC_ADDRESS
   end